requirements.txt
Package                       Version 
----------------------------- --------
asgiref                       3.7.2   
attrs                         23.1.0  
backports.zoneinfo            0.2.1   
Django                        4.2.3   
django-cors-headers           4.2.0   
djangorestframework           3.14.0  
djangorestframework-simplejwt 5.2.2   
drf-spectacular               0.26.4  
importlib-resources           6.0.0   
inflection                    0.5.1   
jsonschema                    4.18.4  
jsonschema-specifications     2023.7.1
pip                           20.0.2  
pkg-resources                 0.0.0   
pkgutil-resolve-name          1.3.10  
PyJWT                         2.8.0   
pytz                          2023.3  
PyYAML                        6.0.1   
referencing                   0.30.0  
rpds-py                       0.9.2   
setuptools                    44.0.0  
sqlparse                      0.4.4   
typing-extensions             4.7.1   
uritemplate                   4.1.1   
zipp                          3.16.2  
